Fixing some missing task > volunteer task renaming
This was causing views not to be located correctly.
