+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/VolunteerTask/Delete.cshtml
@@ -8,7 +8,7 @@
             <li><a asp-controller="Campaign" asp-action="Index" asp-area="Admin">Campaigns</a></li>
             <li><a asp-controller="Campaign" asp-action="Details" asp-route-id="@Model.CampaignId" asp-area="Admin">@Model.CampaignName</a></li>
             <li><a asp-controller="Event" asp-action="Details" asp-route-id="@Model.EventId" asp-area="Admin">@Model.EventName</a></li>
-            <li><a asp-controller="Task" asp-action="Details" asp-route-eventId="@Model.EventId" asp-route-id="@Model.Id" asp-area="Admin">@Model.Name</a></li>
+            <li><a asp-controller="VolunteerTask" asp-action="Details" asp-route-eventId="@Model.EventId" asp-route-id="@Model.Id" asp-area="Admin">@Model.Name</a></li>
             <li>Delete</li>
         </ol>
     </div>
@@ -23,7 +23,7 @@
 </div>
 
 <div class="row">
-    <form asp-controller="Task" asp-action="Delete">
+    <form asp-controller="VolunteerTask" asp-action="Delete">
         <input type="hidden" asp-for="Id" />
         <input type="hidden" asp-for="EventId" />
         <input type="hidden" asp-for="UserIsOrgAdmin" />
@@ -65,7 +65,7 @@
 
             <div class="form-actions no-color">
                 <button type="submit" class="btn btn-default">Delete</button>
-                <a asp-area="Admin" asp-controller="Task" asp-action="Details" asp-route-id="@Model.Id" class="btn btn-default">Cancel</a>
+                <a asp-area="Admin" asp-controller="VolunteerTask" asp-action="Details" asp-route-id="@Model.Id" class="btn btn-default">Cancel</a>
             </div>
         </div>
     </form>
