+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/VolunteerTask/Details.cshtml
@@ -18,7 +18,7 @@
 
 <div class="row">
     <div class="col-md-10">
-        <h2>@Model.Name <a asp-area="Admin" asp-controller="Task" asp-action="Edit" asp-route-id="@Model.Id" class="btn btn-default"><i class="fa fa-edit"></i></a> <a asp-area="Admin" asp-controller="Task" asp-action="Delete" asp-route-id="@Model.Id" class="btn btn-danger"><i class="glyphicon glyphicon-trash"></i></a></h2>
+        <h2>@Model.Name <a asp-area="Admin" asp-controller="VolunteerTask" asp-action="Edit" asp-route-id="@Model.Id" class="btn btn-default"><i class="fa fa-edit"></i></a> <a asp-area="Admin" asp-controller="VolunteerTask" asp-action="Delete" asp-route-id="@Model.Id" class="btn btn-danger"><i class="glyphicon glyphicon-trash"></i></a></h2>
         <p>@Model.Description</p>
         <p>
             <time value="Model.StartDateTime"></time> - <time value="Model.EndDateTime"></time>
@@ -113,7 +113,7 @@
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
                 <h4 class="modal-title" id="messageVolunteersLabel">Message All Volunteers</h4>
             </div>
-            <form asp-area="Admin" asp-controller="Task" asp-action="MessageAllVolunteers" method="post">
+            <form asp-area="Admin" asp-controller="VolunteerTask" asp-action="MessageAllVolunteers" method="post">
                 <div class="modal-body">
                     <input type="hidden" value="@Model.Id" name="volunteerTaskId" id="message-volunteerTaskId">
                     <div class="form-group">
